backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
click==7.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.52
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.9
spleeter==2.4.0
openai-whisper==20231117
ffmpeg-python==0.2.0
python-dotenv==1.0.0
numpy==1.23.5
pandas==1.5.3
jinja2==3.0.3
itsdangerous==2.0.1
demucs
